requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
billiard==3.6.4.0
binaryornot==0.4.4
black==23.3.0
celery==5.2.7
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookiecutter==2.1.1
Django==4.2.1
django-browser-reload==1.8.0
django-extensions==3.2.1
django-filter==23.2
django-tailwind==3.5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fancycompleter==0.9.1
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jinja2-time==0.2.0
kombu==5.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
model-bakery==1.11.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
openai==0.27.6
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
pdbpp==0.10.3
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
Pygments==2.15.1
pyrepl==0.9.0
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wmctrl==0.4
XlsxWriter==3.1.0
yarl==1.9.2
python-dotenv
pyquery==2.0.0